Initial Commit of fully working version
This is the first fully-working version of marvelMailer. It went through a
fair amount of iteration to get here, but it's lost to the winds of time.

Actually, it only took a day.. :P

At the moment, it only practically works on Windows, run interactively.
The next interaction should be more platform-agnostic, and simplify some
of the files.