First attempt for calendar handling
All the functions for actually constructing and adding events to the
calendar are done, but untested because the big issue is authentication.

It's not done. I've tried to use a more future-proof library for
authenticating, but it looks like it's just not ready for use in the
way I would like i.e with stored credentials.

Also updated the README to include some of the new features & clarify
how to get the secrets necessary for authentication.