Cleaned up the code/comments
Made it more pythonic, following pydocs standards etc.

Also removed some hard-coded values.

Also brought the README more up to date. It's still a bit clunky
because the process for acquiring all the tokens and secrets necessary
for calendar authentication and notifier API access is complex, and not
well explained.

Also added an icon, which could be used as a logo or something...