Proactively refresh credentials token
Previously, the credentials token loaded from disk was only refreshed
if it was found to be still valid.

Otherwise it would be replaced, by asking the user to reauthenticate.
This was wholely unneccesary, because the only reason the token was
invalid was because it had expired, and thus needed refreshing.

Tokens are now refreshed after being loaded from disk, and if they are
still invalid or if the refresh failed, the user is asked to
reauthenticate.