Updated README
It's much more accurate now, with a lot more information on how things
work, and how to get started.

The client_secret field in the config file has also been renamed to
client_secret_file, because this is more accurate.

Also, removed the template folder, because it was wildly outdated and
didn't contain anything that wasn't already provided by the README.