Add more error notifications for error situations
Added error notifications if either credentials (cookie or calendar
credentials) fail.

Also added a check for if the 'My Account' page loaded correctly & for
if the active subscriptions table is on the page. These are separate
checks because it's possible to successfully log in, but not have any
active subscriptions to scrape.

Also renamed notifier object to 'notifier', because 'n' was meaningless.