Added option for ETA events
Added an option to the config file that can be specified to add events
to the calendar for an expected delivery date of shipped issues.
This meant tweaking the event adding code to allow specifying a date.
Also added this information to the README.

Also a bit of a general cleanup of the README + added instructions for
first run setup.